✨ feat #64, refactor: Large 사이즈인 경우 가격 500원 추가하기
- MenuItem 별도 컴포넌트로 분리: 메인 영역, 모달 영역, 카트 영역에서 모두 사용되는 공통 컴포넌트이므로
- MenuItem 생성되는 위치마다 다르게 스타일 적용하기
- OrderModal 하위 컴포넌트 모듈화
-
