+++ fe/src/components/Main/Cart.module.css
@@ -1,0 +1,149 @@
+.orderItem {
+  display: flex;
+  flex-direction: column;
+  justify-content: space-evenly;
+  align-items: center;
+  width: 6rem;
+  height: 6rem;
+  font-family: var(--font-family);
+  font-weight: var(--font-weight-medium);
+  font-size: 0.8rem;
+  border: 1px solid var(--color-primary);
+  border-radius: 1rem;
+  background-color: var(--color-white);
+}
+
+.orderItem img {
+  width: 40%;
+  height: 40%;
+  border-radius: 50%;
+}
+
+.orderItem > span {
+  width: 100%;
+  text-align: center;
+}
+
+.orderItem:hover {
+  cursor: default;
+  transform: none;
+}
+
+.cart {
+  display: flex;
+  background-color: var(--color-off-white);
+  justify-content: space-between;
+  padding: 1rem;
+  width: 100%;
+  height: 20rem;
+  border-top: 2px solid var(--color-primary);
+}
+
+.orderItems {
+  width: 50%;
+  height: 100%;
+  overflow-y: scroll;
+  &::-webkit-scrollbar {
+    display: none;
+  }
+  display: flex;
+  justify-content: center;
+  align-items: center;
+  max-width: 40%;
+  flex-wrap: wrap;
+  padding: 1rem;
+  gap: 1.5rem;
+}
+
+.buttons {
+  display: flex;
+  flex-direction: column;
+  justify-content: flex-end;
+  align-items: flex-end;
+  width: calc(100% - 40%);
+  gap: 1rem;
+  padding: 0 1rem;
+}
+
+.menuCancelButton {
+  font-weight: var(--font-weight-bold);
+  color: var(--color-white);
+  width: 1.5rem;
+  height: 1.5rem;
+  padding: 0.5rem;
+  background-color: var(--color-black);
+  border-radius: 50%;
+  position: absolute;
+  display: flex;
+  justify-content: center;
+  align-items: center;
+  z-index: 1;
+  top: -0.5rem;
+  right: -0.5rem;
+}
+
+.allCancelButton {
+  padding: 1rem 5rem;
+  border-radius: 0.5rem;
+  background-color: var(--color-black);
+  display: flex;
+  justify-content: center;
+  align-items: center;
+  font-size: 1.5rem;
+  color: var(--color-white);
+  font-weight: var(--font-weight-bold);
+  cursor: pointer;
+}
+
+.orderButton {
+  padding: 1rem 5rem;
+  background-color: var(--color-primary);
+  color: var(--color-white);
+  font-family: var(--font-family);
+  font-weight: var(--font-weight-bold);
+  font-size: 1.5rem;
+  border: none;
+  border-radius: 0.5rem;
+  cursor: pointer;
+}
+
+.timer {
+  font-family: var(--font-family);
+  font-weight: var(--font-weight-bold);
+  font-size: 1rem;
+  color: var(--color-primary);
+}
+
+.menuCancelButton:hover {
+  cursor: pointer;
+  transform: scale(1.1);
+  transition: all 0.3s ease-in-out;
+}
+
+.totalPrice {
+  font-family: var(--font-family);
+  font-weight: var(--font-weight-bold);
+  font-size: 1.5rem;
+  color: var(--color-primary);
+}
+
+.amount {
+  font-weight: var(--font-weight-bold);
+  color: var(--color-white);
+  width: 1.5rem;
+  height: 1.5rem;
+  padding: 0.5rem;
+  background-color: var(--color-primary);
+  border-radius: 50%;
+  position: absolute;
+  display: flex;
+  justify-content: center;
+  align-items: center;
+  z-index: 1;
+  top: -0.5rem;
+  left: -0.5rem;
+}
+
+.itemWrapper {
+  position: relative;
+}
