+++ fe/src/components/Main/Main.module.css
@@ -4,7 +4,6 @@
   width: 100%;
   height: 100%;
   overflow-y: scroll;
-  padding: 2rem;
   display: flex;
   flex-wrap: wrap;
   flex-grow: 1;
@@ -15,193 +14,15 @@
   }
 }
 
+/* TODO: height를 어떻게 설정해야 main이 스크롤 될 때도 아래까지 다 덮을 수 있을까? */
 .mainDim {
-  height: 100vh;
-}
-
-.menuItem {
-  position: relative;
-  width: 18rem;
-  height: 18rem;
-  gap: 0.5rem;
+  position: absolute;
   display: flex;
-  flex-direction: column;
   justify-content: center;
   align-items: center;
-  margin: 1rem;
-  padding: 2rem;
-  border: 1px solid black;
-  border-radius: 1rem;
-  background-color: var(--color-off-white);
-  box-shadow: 0 0 10px rgba(0, 0, 0, 0.5);
-  transition: all 0.3s ease-in-out;
-}
-
-.menuItem:hover {
-  cursor: pointer;
-  transform: scale(1.05);
-}
-
-.menuItem img {
-  border-radius: 1rem;
-  object-fit: scale-down;
-}
-
-.cart {
-  display: flex;
-  background-color: var(--color-off-white);
-  justify-content: space-between;
-  padding: 1rem;
   width: 100%;
-  height: 20rem;
-  border-top: 2px solid var(--color-primary);
-}
-
-.orderItems {
-  width: 50%;
-  height: 100%;
-  overflow-y: scroll;
-  &::-webkit-scrollbar {
-    display: none;
-  }
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  max-width: 40%;
-  flex-wrap: wrap;
-  padding: 1rem;
-  gap: 1.5rem;
-}
-
-.orderItem {
-  width: 6rem;
-  height: 6rem;
-  margin: 0;
-  padding: 0.8rem;
-  font-family: var(--font-family);
-  font-weight: var(--font-weight-medium);
-  font-size: 10px;
-  line-height: auto;
+  height: 100vh;
   overflow: hidden;
-  white-space: wrap;
-}
-
-.orderItem img {
-  width: 80%;
-  height: 80%;
-  object-fit: scale-down;
-  border-radius: 0.5rem;
-}
-
-.orderItem:hover {
-  cursor: default;
-  transform: none;
-}
-
-.itemWrapper {
-  position: relative;
-}
-
-.amount {
-  font-weight: var(--font-weight-bold);
-  color: var(--color-white);
-  width: 1.5rem;
-  height: 1.5rem;
-  padding: 0.5rem;
-  background-color: var(--color-primary);
-  border-radius: 50%;
-  position: absolute;
-  display: flex;
-  justify-content: center;
-  align-items: center;
+  background-color: rgba(128, 128, 128, 0.7);
   z-index: 1;
-  top: -0.5rem;
-  left: -0.5rem;
-}
-
-.buttons {
-  display: flex;
-  flex-direction: column;
-  justify-content: flex-end;
-  align-items: flex-end;
-  width: calc(100% - 40%);
-  gap: 1rem;
-  padding: 0 1rem;
-}
-
-.menuCancelButton {
-  font-weight: var(--font-weight-bold);
-  color: var(--color-white);
-  width: 1.5rem;
-  height: 1.5rem;
-  padding: 0.5rem;
-  background-color: var(--color-black);
-  border-radius: 50%;
-  position: absolute;
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  z-index: 1;
-  top: -0.5rem;
-  right: -0.5rem;
-}
-
-.allCancelButton {
-  padding: 1rem 5rem;
-  border-radius: 0.5rem;
-  background-color: var(--color-black);
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  font-size: 1.5rem;
-  color: var(--color-white);
-  font-weight: var(--font-weight-bold);
-  cursor: pointer;
-}
-
-.orderButton {
-  padding: 1rem 5rem;
-  background-color: var(--color-primary);
-  color: var(--color-white);
-  font-family: var(--font-family);
-  font-weight: var(--font-weight-bold);
-  font-size: 1.5rem;
-  border: none;
-  border-radius: 0.5rem;
-  cursor: pointer;
-}
-
-.timer {
-  font-family: var(--font-family);
-  font-weight: var(--font-weight-bold);
-  font-size: 1rem;
-  color: var(--color-primary);
-}
-
-.menuCancelButton:hover {
-  cursor: pointer;
-  transform: scale(1.1);
-  transition: all 0.3s ease-in-out;
-}
-
-.best {
-  position: absolute;
-  top: -1rem;
-  left: -2rem;
-  width: 4rem;
-  height: 4rem;
-  background-image: url('../../../public/assets/icon/best.png');
-  background-size: cover;
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  color: var(--color-white);
-  font-weight: var(--font-weight-bold);
-}
-
-.totalPrice {
-  font-family: var(--font-family);
-  font-weight: var(--font-weight-bold);
-  font-size: 1.5rem;
-  color: var(--color-primary);
 }
