+++ fe/src/components/MenuItem/MenuItem.module.css
@@ -1,0 +1,42 @@
+.menuItem {
+  position: relative;
+  width: 18rem;
+  height: 18rem;
+  gap: 0.5rem;
+  display: flex;
+  flex-direction: column;
+  justify-content: center;
+  align-items: center;
+  margin: 1rem;
+  padding: 2rem;
+  border: 1px solid black;
+  border-radius: 1rem;
+  background-color: var(--color-off-white);
+  box-shadow: 0 0 10px rgba(0, 0, 0, 0.5);
+  transition: all 0.3s ease-in-out;
+}
+
+.menuItem:hover {
+  cursor: pointer;
+  transform: scale(1.05);
+}
+
+.menuItem img {
+  border-radius: 1rem;
+  object-fit: scale-down;
+}
+
+.best {
+  position: absolute;
+  top: -1rem;
+  left: -2rem;
+  width: 4rem;
+  height: 4rem;
+  background-image: url('../../../public/assets/icon/best.png');
+  background-size: cover;
+  display: flex;
+  justify-content: center;
+  align-items: center;
+  color: var(--color-white);
+  font-weight: var(--font-weight-bold);
+}
