--- fe/src/components/Modal/OrderModal.tsx
@@ -1,220 +1,0 @@
-import MenuItem from 'components/Main/MenuItem';
-import Button from 'components/atoms/Button';
-import { ProductInfo, ProductOrder } from 'pages/types';
-import { useState } from 'react';
-import ModalStyles from './Modal.module.css';
-import styles from './OrderModal.module.css';
-
-interface OrderModalProps {
-  menu: ProductInfo;
-  closeOrderModal: () => void;
-  handleAddOrder: (menuOrder: ProductOrder) => void;
-}
-
-export default function OrderModal({ menu, closeOrderModal, handleAddOrder }: OrderModalProps) {
-  const [selectedSize, setSelectedSize] = useState('');
-  const [selectedTemp, setSelectedTemp] = useState('');
-  const [amount, setAmount] = useState(1);
-  const sendOrderData = (handleAddOrder: (menuOrder: ProductOrder) => void) => {
-    const orderData = {
-      productId: menu.productId,
-      name: menu.name,
-      size: selectedSize,
-      temperature: selectedTemp,
-      amount: amount,
-    };
-    handleAddOrder(orderData);
-  };
-
-  return (
-    <div className={styles.wrap}>
-      <CloseButton closeOrderModal={closeOrderModal} />
-      <div className={styles.contents}>
-        <MenuItem
-          productId={menu.productId}
-          menuName={menu.name}
-          menuImg={menu.imgUrl}
-          menuPrice={menu.price}
-          hasHot={menu.hasHot}
-          hasIce={menu.hasIce}
-          hasLarge={menu.hasLarge}
-          hasSmall={menu.hasSmall}
-        />
-        <MenuOption
-          hasLarge={menu.hasLarge}
-          hasSmall={menu.hasSmall}
-          hasHot={menu.hasHot}
-          hasIce={menu.hasIce}
-          selectedSize={selectedSize}
-          selectedTemp={selectedTemp}
-          setSelectedSize={setSelectedSize}
-          setSelectedTemp={setSelectedTemp}
-          amount={amount}
-          setAmount={setAmount}
-        />
-      </div>
-      <AddButton
-        handleAddOrder={handleAddOrder}
-        closeOrderModal={closeOrderModal}
-        sendOrderData={sendOrderData}
-        selectedSize={selectedSize}
-        selectedTemp={selectedTemp}
-      />
-    </div>
-  );
-}
-
-function CloseButton({ closeOrderModal }: { closeOrderModal: () => void }) {
-  return (
-    <button className={ModalStyles.closeButton} type="button" onClick={closeOrderModal}>
-      X
-    </button>
-  );
-}
-
-interface MenuOptionProps extends SizeOptionProps, TempOptionProps, AmountCounterProps {
-  setSelectedSize: (size: string) => void;
-  setSelectedTemp: (temp: string) => void;
-}
-
-function MenuOption({
-  hasLarge,
-  hasSmall,
-  hasHot,
-  hasIce,
-  selectedSize,
-  setSelectedSize,
-  selectedTemp,
-  setSelectedTemp,
-  amount,
-  setAmount,
-}: MenuOptionProps) {
-  return (
-    <div className={styles.optionWrap}>
-      <SizeOption
-        hasSmall={hasSmall}
-        hasLarge={hasLarge}
-        selectedSize={selectedSize}
-        setSelectedSize={setSelectedSize}
-      />
-      <TempOption hasHot={hasHot} hasIce={hasIce} selectedTemp={selectedTemp} setSelectedTemp={setSelectedTemp} />
-      <AmountCounter amount={amount} setAmount={setAmount} />
-    </div>
-  );
-}
-
-interface SizeOptionProps {
-  hasLarge: boolean;
-  hasSmall: boolean;
-  selectedSize: string;
-  setSelectedSize: (size: string) => void;
-}
-
-function SizeOption({ hasSmall, hasLarge, selectedSize, setSelectedSize }: SizeOptionProps) {
-  return (
-    <div className={hasSmall && hasLarge ? styles.dualButtonWrap : styles.singleButtonWrap}>
-      {hasSmall && (
-        <Button
-          label={'Small'}
-          className={styles.optionButton}
-          isSelected={selectedSize === 'Small'}
-          selectedClassName={styles.selected}
-          onClick={() => setSelectedSize('Small')}
-        />
-      )}
-      {hasLarge && (
-        <Button
-          label={'Large'}
-          className={styles.optionButton}
-          isSelected={selectedSize === 'Large'}
-          selectedClassName={styles.selected}
-          onClick={() => setSelectedSize('Large')}
-        />
-      )}
-    </div>
-  );
-}
-
-interface TempOptionProps {
-  hasHot: boolean;
-  hasIce: boolean;
-  selectedTemp: string;
-  setSelectedTemp: (temp: string) => void;
-}
-
-function TempOption({ hasHot, hasIce, selectedTemp, setSelectedTemp }: TempOptionProps) {
-  return (
-    <div className={hasHot && hasIce ? styles.dualButtonWrap : styles.singleButtonWrap}>
-      {hasHot && (
-        <Button
-          label={'Hot'}
-          className={styles.optionButton}
-          isSelected={selectedTemp === 'Hot'}
-          selectedClassName={styles.selected}
-          onClick={() => setSelectedTemp('Hot')}
-        />
-      )}
-      {hasIce && (
-        <Button
-          label={'Ice'}
-          className={styles.optionButton}
-          isSelected={selectedTemp === 'Ice'}
-          selectedClassName={styles.selected}
-          onClick={() => setSelectedTemp('Ice')}
-        />
-      )}
-    </div>
-  );
-}
-
-interface AmountCounterProps {
-  amount: number;
-  setAmount: (amount: number) => void;
-}
-
-function AmountCounter({ amount, setAmount }: AmountCounterProps) {
-  const plusAmount = () => {
-    amount === 99 ? setAmount(99) : setAmount(amount + 1);
-  };
-
-  const minusAmount = () => {
-    amount === 1 ? setAmount(1) : setAmount(amount - 1);
-  };
-
-  return (
-    <div className={styles.counter}>
-      <button onClick={minusAmount}>-</button>
-      <div>
-        <span>{amount}</span>
-      </div>
-      <button onClick={plusAmount}>+</button>
-    </div>
-  );
-}
-
-interface AddButtonProps {
-  closeOrderModal: () => void;
-  sendOrderData: (handleAddOrder: (menuOrder: ProductOrder) => void) => void;
-  handleAddOrder: (menuOrder: ProductOrder) => void;
-  selectedSize: string;
-  selectedTemp: string;
-}
-
-function AddButton({ closeOrderModal, sendOrderData, handleAddOrder, selectedSize, selectedTemp }: AddButtonProps) {
-  const isDisabled = selectedSize === '' || selectedTemp === '';
-  const handleAddButtonClick = () => {
-    closeOrderModal();
-    sendOrderData(handleAddOrder);
-  };
-
-  return (
-    <button
-      className={`${styles.addButton} ${styles.disabled}`}
-      type="button"
-      disabled={isDisabled}
-      onClick={handleAddButtonClick}
-    >
-      담기
-    </button>
-  );
-}
