+++ fe/src/components/Modal/OrderModal/index.tsx
@@ -1,0 +1,3 @@
+import OrderModal from './OrderModal';
+
+export default OrderModal;
