+++ fe/src/constant.ts
@@ -1,0 +1,1 @@
+export const EXTRA_PRICE = 500;
